Update "better_stocks", show item counts and count build items as "red"
This involves hooking into two in-game functions, which is ugly but
easier than rewriting them within the plugin. Also, since this is the only
place that uses them, I'm not going to bother adding them to symbols.xml.

Also update "more_jobs" plugin to use a better "one-time init" check for
the createquota viewscreen which requires one less hook.